+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Media can be renamed, with each article that uses the media automatically renaming its references as well.
 * You can select a given image from the media gallery as the cover image for a wiki. Go to the image's edit page and select "Use this image as the wiki's cover image."
 * Error pages for wikis, media, etc. not found are much cleaner now.
+* Titles for many pages are more explicit.
 
 ### Bugfixes